pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bizon"
version = "0.1.2"
description = "Extract and load your data reliably from API Clients with native fault-tolerant and checkpointing mechanism."
authors = [
    { name = "Antoine Balliet", email = "antoine.balliet@gmail.com" },
    { name = "Anas El Mhamdi", email = "anas.elmhamdi@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "requests>=2.28.2",
    "pyyaml>=6.0.1",
    "pydantic>=2.8.2",
    "backoff>=2.2.1",
    "pendulum>=3.0.0",
    "dpath>=2.2.0",
    "pydantic-extra-types>=2.9.0",
    "loguru>=0.7.2",
    "pyarrow>=16.1.0",
    "google-cloud-storage>=2.17.0",
    "sqlalchemy>=2.0.32",
    "polars>=1.16.0",
    "click>=8.1.7",
    "pytz>=2024.2",
    "tenacity>=9.0.0",
    "simplejson>=3.20.1",
    "orjson>=3.10.16",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.9"]
bigquery = [
    "google-cloud-bigquery-storage>=2.25.0",
    "google-cloud-bigquery>=3.25.0",
    "sqlalchemy-bigquery>=1.11.0",
    "protobuf>=4.24.0",
]
kafka = [
    "confluent-kafka>=2.6.0",
    "fastavro>=1.9.7",
    "avro>=1.12.0",
    "kafka-python>=2.0.2",
]
rabbitmq = ["pika>=1.3.2"]
gsheets = ["gspread>=6.1.2"]
datadog = ["datadog>=0.50.2", "ddtrace>=3.10.0"]

[project.scripts]
bizon = "bizon.cli.main:cli"

[dependency-groups]
test = [
    "pytest>=7.2.2,<8",
    "requests-mock>=1.10.0",
    "python-dotenv>=1.0.1",
    "pytest-lazy-fixture>=0.6.3",
    "faker>=26.0.0",
]
dev = [
    "ipykernel>=6.29.5",
    "snakeviz>=2.1.2",
    "yappi>=1.3.2",
    "pre-commit>=3.8.0",
    "ruff>=0.8.4",
]

[tool.hatch.build.targets.wheel]
packages = ["bizon"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E711",   # comparison to None (existing code)
    "E712",   # comparison to True/False (common in tests)
    "E721",   # type comparison using == (existing code)
    "F841",   # unused variable (common in tests with pytest.raises)
    "B006",   # mutable default arguments
    "B007",   # unused loop variable
    "B008",   # function calls in argument defaults
    "B015",   # pointless comparison (existing code)
    "B019",   # lru_cache on methods
    "B023",   # function binding loop variable
    "B026",   # star-arg unpacking after keyword
    "B027",   # empty method in abstract class
    "B904",   # raise from in except (existing code)
    "C4",     # flake8-comprehensions (existing code style)
    "UP006",  # deprecated typing (List -> list) - existing code
    "UP035",  # deprecated typing imports - existing code
]

[tool.ruff.lint.isort]
known-first-party = ["bizon"]
